# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface StoredExperience {
  id: string;
  userId: string;
  company: string;
  role: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  isCurrent?: boolean;
  /** @nullable */
  description?: string | null;
  skills?: string[];
  createdAt: string;
  updatedAt: string;
}

export interface ExperienceInput {
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  role: string;
  city?: string;
  startDate?: string;
  endDate?: string;
  isCurrent?: boolean;
  description?: string;
  skills?: string[];
}

export interface ExperienceListEnvelope {
  experiences: StoredExperience[];
}

export interface ExperienceEnvelope {
  experience: StoredExperience;
}

export const DeleteSuccessValue = {
  success: true,
} as const;
export type DeleteSuccess = typeof DeleteSuccessValue;

export type StoredHighlightType = typeof StoredHighlightType[keyof typeof StoredHighlightType];


export const StoredHighlightType = {
  volunteering: 'volunteering',
  honor: 'honor',
  project: 'project',
  other: 'other',
} as const;

export interface StoredHighlight {
  id: string;
  userId: string;
  type: StoredHighlightType;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  link?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type HighlightInputType = typeof HighlightInputType[keyof typeof HighlightInputType];


export const HighlightInputType = {
  volunteering: 'volunteering',
  honor: 'honor',
  project: 'project',
  other: 'other',
} as const;

export interface HighlightInput {
  type: HighlightInputType;
  /** @minLength 1 */
  title: string;
  description?: string;
  date?: string;
  link?: string;
}

export interface HighlightListEnvelope {
  highlights: StoredHighlight[];
}

export interface HighlightEnvelope {
  highlight: StoredHighlight;
}

export type PublicProfileSectionKey = typeof PublicProfileSectionKey[keyof typeof PublicProfileSectionKey];


export const PublicProfileSectionKey = {
  experiences: 'experiences',
  education: 'education',
  languages: 'languages',
  skills: 'skills',
  highlights: 'highlights',
} as const;

export interface PublicProfileSection {
  key: PublicProfileSectionKey;
  visible: boolean;
  order: number;
}

export interface StoredPublicProfile {
  id: string;
  userId: string;
  slug: string;
  published: boolean;
  /** @nullable */
  photo?: string | null;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  bio?: string | null;
  selectedExperienceIds?: string[];
  sections?: PublicProfileSection[];
  createdAt: string;
  updatedAt: string;
  /**
     * Absolute /p/:slug URL, only present when published.
     * @nullable
     */
  publicUrl: string | null;
}

export interface PublicProfileInput {
  photo?: string;
  headline?: string;
  bio?: string;
  selectedExperienceIds?: string[];
  sections?: PublicProfileSection[];
}

export interface PublicProfileEnvelope {
  profile?: StoredPublicProfile | null;
}

export interface FetchJobRequest {
  /** @minLength 1 */
  url: string;
}

export interface FetchJobResponse {
  text: string;
}

export interface TailorCvRequest {
  /** @minLength 50 */
  jobDescription: string;
  /** Optional list of archived experience IDs to use. If omitted, all saved experiences are considered. */
  experienceIds?: string[];
}

export interface TailoredCVExperience {
  id: string;
  company: string;
  role: string;
  city: string;
  from: string;
  to: string;
  desc: string;
}

export type TailoredCVDataEducationItem = { [key: string]: unknown };

export type TailoredCVDataLanguagesItem = { [key: string]: unknown };

export interface TailoredCVData {
  firstName: string;
  lastName: string;
  title: string;
  email: string;
  phone: string;
  city: string;
  linkedin: string;
  summary: string;
  experiences: TailoredCVExperience[];
  education: TailoredCVDataEducationItem[];
  skills: string[];
  languages: TailoredCVDataLanguagesItem[];
}

export interface TailorCvEnvelope {
  cvData: TailoredCVData;
}

export interface ConfirmTailorCvRequest {
  cvData: TailoredCVData;
  /** @minLength 1 */
  jobDescription: string;
}

export interface ConfirmTailorCvEnvelope {
  /** ID of the saved tailored CV record. */
  savedCvId: string;
}

export interface StoredTailoredCv {
  id: string;
  userId: string;
  jobTitle: string;
  jobDescription: string;
  cvData: TailoredCVData;
  createdAt: string;
}

export interface TailoredCvListEnvelope {
  tailoredCvs: StoredTailoredCv[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};