# lib/api-zod/src/generated/types/highlightInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HighlightInputType } from './highlightInputType';

export interface HighlightInput {
  type: HighlightInputType;
  /** @minLength 1 */
  title: string;
  description?: string;
  date?: string;
  link?: string;
}