# lib/api-zod/src/generated/types/publicProfileEnvelope.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoredPublicProfile } from './storedPublicProfile';

export interface PublicProfileEnvelope {
  profile?: StoredPublicProfile | null;
}