# lib/api-zod/src/generated/types/publicProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicProfileSection } from './publicProfileSection';

export interface PublicProfileInput {
  photo?: string;
  headline?: string;
  bio?: string;
  selectedExperienceIds?: string[];
  sections?: PublicProfileSection[];
}