# lib/api-zod/src/generated/types/storedPublicProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicProfileSection } from './publicProfileSection';

export interface StoredPublicProfile {
  id: string;
  userId: string;
  slug: string;
  published: boolean;
  /** @nullable */
  photo?: string | null;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  bio?: string | null;
  selectedExperienceIds?: string[];
  sections?: PublicProfileSection[];
  createdAt: Date;
  updatedAt: Date;
  /**
     * Absolute /p/:slug URL, only present when published.
     * @nullable
     */
  publicUrl: string | null;
}